Improve cf service injection
